backend/requirements.txt
# FastAPI & Backend Dependencies
fastapi
uvicorn
python-multipart
pydantic

# Machine Learning & Data Processing
torch
torchvision
numpy
scikit-learn
matplotlib
seaborn
ijson
psutil
ndjson

# Image Processing
pillow

# Utilities
python-dotenv

# Frontend Dependencies (install via npm in frontend directory)
# See frontend/package.json for React, TypeScript, and other frontend packages
